Auto resize columns
The table's rows will automatically resize
1. on load
2. on window resize
3. when adding/removing a row

A nice feature to add in the future would be to avoid automatically resize columns that have been manually adjusted by the user. This is not very straightforward to implement so it will have to be revisited in a future cl.


Change-Id: I7d50461e3a635e80cbc1db3530711a44e8421b8b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6088932
Commit-Queue: Pietro Scuttari <pietroscutta@google.com>
Reviewed-by: Vaghinak Vardanyan <vaghinak@google.com>