[auth] Drop default port number from hostnames in ID token "aud".
On the client side, when using "https://${host}" audience, we now
always drop ":443" from the host if it is present (other non-standard
ports are preserved). This is needed because Cloud Run doesn't like
the presence of the port there.

On the server side, we now normalize hostnames before comparing
them by dropping default port numbers. This is needed for
compatibility with existing clients that still may send "...:443"
explicitly either in "aud" or in "Host" header.

R=gregorynisbet@google.com

Change-Id: Ie26806cd2259ee96ace983bd99bef3cd054a7eac
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6154976
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>