[swarming] Add a mechanism to update BotInfo and record BotEvents.
The conceptual model here is that BotInfo is always updated by
sending it "a message" in a form of BotInfoUpdate. BotInfoUpdate
contains what parts of BotInfo should be updated.

Unlike the Python implementation, this mechanism will be the only
way BotInfo can be updated and it will guarantee all updates and
BotEvent recording happen transactionally. This will eliminate
a class of bugs where bot events and bot info contradict one a
another.

This may require some benchmarking and/or retry tweaking to
ensure there are no performance or availability regressions
(there should be no impact on costs, since it seems transactions
are "free" and we pay only for actual reads or writes).

This also adds a best effort mechanism to skip duplicate events.
It only skips the last recorded event. It should be good enough
for the most common case of the bot just retrying the RPC call
until it succeeds.

R=chanli@chromium.org
BUG=b/355013282

Change-Id: Id3f72798fe1955a51d9b687d1cdfac5f62bd6a83
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6161754
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>