[swarming] Add a helper library to work with bot state JSON dict.
It deserializes JSON lazily and allows to read only specific keys:
Swarming server reads very few keys from the state, deserializing
all of it all the time is a waste of time.

This is used in the bot scanner cron to avoid deserializing the
state multiple times and it will also be used in /bot/handshake
handler when parsing the state dict received from the bot.

R=chanli@chromium.org
BUG=b/355013282

Change-Id: I2f9c21ac2855dc7a3e9838ce4b21076ff56d6ef9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6174305
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>