[cipd] Remove defunct debugging metric.
This metric only approximately reported the data it purported to,
and has an old configuration which only applies to a subset of
areas where this might be useful.

Tear it out for now - we can re-add a better version of this
trivially later if needed.

R=fancl, vadimsh

Change-Id: Ic1aacf81028acc01e8bf5a3bd01c973fbb98fad1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6169824
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>