[cipd] Improve retry metrics to be clearer.
This replaces the retry distribution with two counters - one for
each retry, and one per ReadAt call.

The combination of these should paint a very clear picture about
when retries are happening, and should allow us to clearly see the
number of retries which are occurring on a per-call basis.

R=bryner, ddoman, fancl, vadimsh

Bug: 40202444,361711863
Change-Id: Ie0e4f23505e24e68df8ccddbc91cde923558015e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6191850
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Scott Lee <ddoman@chromium.org>