[swarming] Tweak BotInfoUpdate API a little bit.
A full *BotGroup is not always available. In particular it is not
available in a scenario where a bot is removed from the config
while it still runs a task. We use the config "frozen" in the
session in that case. But the session doesn't carry all
information that exists in *BotGroup, which makes using
BotInfoUpdate problematic.

While we can theoretically put more stuff into the session,
in this particular case we don't really need all of *BotGroup
in BotInfoUpdate. We just need dimensions configured in bots.cfg.
Make BotInfoUpdate accept them directly. In cases where we rely
on the "frozen" session state, they actually won't be needed
at all (because we'll be populating `Dimensions` instead).

Also add a panic in case `Dimensions` don't have the bot ID set,
since this would be a pretty bad violation if this sneaks in.

R=chanli@chromium.org, iannucci@chromium.org
BUG=b/355012930

Change-Id: Ica62e48ab3657797f2b6c850daefe5b3f7568857
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6191416
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>