[swarming] Refactor BotInfoUpdate to allow reusing transaction logic.
There are couple of cases when we need to submit a BotInfo update
transactionally with some other datastore work. I initially planned
to allow callers to run their own transaction and make BotInfo
update be a part of it. But it would actually be better to drive
the transaction via BotInfoUpdate.Submit(...), since it has
configured retries and monitoring.

So now BotInfoUpdate can have an optional callback that can execute
any extra transactional work and/or skip the update entirely (if
the conditional that triggered the update no longer holds).

This will be used in the cron job that marks bots as dead.

R=iannucci@chromium.org
BUG=b/355013354

Change-Id: I9d6d9d90bc7112728a8d3323fe74c29637303a7d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6216822
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>