[swarming] Use BotInfoUpdate mechanism when deleting a bot.
There are two advantages:
1. We'll have an event recorded.
2. BotInfoUpdate becomes the only place that deals with BotInfo changes
   and it can implement some invariants (like making sure if a bot
   disappears the task it was running is marked as abandoned).

Note that we also delete BotInfo entities via Cloud Datastore TTL and
this will not record any event nor keep any invariants. But for the TTL
mechanism to kick in, a bot should be in dead state for a while. Such
bots will have BotEventMissing event recorded, and this is where we
will enforce invariants as well. BotEventDeleted is for cases when
a recently alive bot is quickly deleted via an RPC (such bots will
never see BotEventMissing event).

R=iannucci@chromium.org
BUG=b/355013593

Change-Id: Icd04241a53253a77e723cf196ec37566c6c74497
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6217309
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>