Fleet Console: Add support for 3-dot menu in column headers to enable more table column options
Screenshot: https://screenshot.googleplex.com/9b44eNmdXkhcGL3

This is a low hanging fruit improvement because we can use what MUI provides for us out of the box. There's few downsides to adding this because users can still click on the table header to sort.

Later, we can make additional improvements to make this header more useful, but in the short-term it adds an "Unsort" option which was previously not available.

TEST=Manual. Opted not to add automated tests for this because the cost of setting up a grid context for testing might be high, and the amount of customizations here are small.

Bug:393602662
Change-Id: I539fc80bc5d99324e3d0093269bec56e11d9f643
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6220274
Reviewed-by: Gowri Denduluri <gowriden@google.com>
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>