[protowalk] Refactor to remove all locks.
This changes the interface of the protowalk library so that the
caller directly constructs a Walker which is keyed to a specific
proto Message, and has a specific set of FieldProcessors.

The caller can then call `Walker.Execute(msg)` to evaluate an
instance of the Message into a Result.

This has a type-safe Walker[M] and a type-checked DynamicWalker[M]
which will be useful in dynamic scenarios (for example - when
constructing Walkers for luci server applications over their
registered gRPC interfaces).

This also removes the need to pre-register FieldProcessor instances,
instead they are just directly provided during the construction of
the Walker, and include `ShouldProcess` as part of the interface.

R=chanli, vadimsh

Change-Id: I0b37bfb5ce9b9e1f90ae265783d6ff8646eefc1b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5871428
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>