[logging] Refactor logging lib to allow passing more data.
This also makes it slightly more efficient by using a single
context key instead of 3 context keys (and by reducing context
lookups per log line to 1 instead of 3).

I want to be able to attach stack traces to selected log messages,
to later read them in logging implementations that understand stack
traces (Cloud Error Reporting).

The only way to do it now is either to shove it into untyped
logging.Fields (and make sure to ignore it in loggers that don't
know what to do with stack traces), or introduce yet another bit
of state in the context with a separate context key.

Instead we can group all the context state in a single structs
(called LogContext here) and later just add a new typed field to
this struct.

R=iannucci@chromium.org, gregorynisbet@google.com

Change-Id: I2e526dc11723a29b17f08f70349f666e23ba7532
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6221587
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>