[testclock] Reimplement clearPendingTimer.
The previous implementation seems to be treating pendingTimers as
both a list and a heap, and it's not clear that this is generally
safe to do. On top of this it has an obvious bug (using the 0 index
instead of `i`).

This implementation simply re-heapifies after the cancellation and
removal of affected timers.

R=vadimsh

Bug: 353565742
Change-Id: I4a5a9191e992ef4decfb7b7c89a64c3019452b32
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6227030
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>