Improve filters responsiveness.
Added virtualization for filter options.

You can see the difference in the videos (tested on company chromebook in dev build with 1000 devices):

Without virtualization:
https://screencast.googleplex.com/cast/NjczNDIwNDEwNzc1MTQyNHxjN2FkOGY4NC1jYw

With virtualization:
https://screencast.googleplex.com/cast/NTQzNDgzMzU1ODgzMTEwNHw2MmIwMjIxMi02NQ

Minor thing to notice - when using virtualization we don't know what the width of the widest list item will be, so we need to set it. Also currently column picker reuses the same component as selected chip, so for now I set the same width for both of them (in future it would probably be better to split them, since column picker doesn't need virtualization)

Bug=b:391554425

Change-Id: Ib2f2bf72f862e9b9c0a2edc566957ffeeb2e3824
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6233833
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Konrad Polarczyk <polarczyk@google.com>
Reviewed-by: Pietro Scuttari <pietroscutta@google.com>