[testing] Local-only tests!
Right now we^H^HI kick flaky tests out of the tree by skipping them.

Skipping tests is not very nice, making them local-only is a much milder
response than skipping them.

This does make it possible for a test to live for a long time as a
local-only test and become a load-bearing local-only test.

It might be worth investigating this as a possibility though.

BUG=None
TEST=also None

Change-Id: I58c58525f9ef4233945023865d97a98a25833146
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6236706
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>