[swarming] Resubmit reservations failed with "bot is dead" status.
This should be safe as long as we use a new reservation ID (otherwise
RBE will just dedup the reservation to the failed one) and if we do
that before the bot has claimed the TaskToRun.

Note that we allow the following scenario:
1. An attempt #0 is submitted to RBE.
2. Bot "A" picks it up.
3. Bot "A" gets stuck somewhere for some time.
4. RBE expires the reservation with "bot is dead" error.
5. An attempt #1 is submitted to RBE.
6. Bot "A" suddenly gets unstuck and claims TaskToRun for execution.
7. Bot "B" picks up the attempt #1 reservation.
8. Bot "B" sees the TaskToRun is already claimed and does nothing.

This should be rare though. If it is problematic for some reason,
we can later add a new parameter to /bot/claim to allow a bot to
claim TaskToRun only if it still "tracking" the RBE reservation
the bot got. In that case in the above scenario, bot "A" would
do nothing and bot "B" would run the task.

R=chanli@chromium.org
BUG=b/392668963

Change-Id: I0e637ce0bd58f9f5bb77ca89fa2b760080e555de
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6250239
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>