[gcep] Fix VM terminating logic in gce-provider
Bug: 395973590

This patch updates the logic to terminate the VMs in gcep. The old logic
ensured that the VMs would be killed if the swarming reported that the
bot was dead (IsDead)*. This causes issues if the VM is rebooting and
swarming lists it as dead. The new flow takes advantage of the field
termination_task_id in swarming bot field to address a few corner cases.
The new logic is as follows
        1. Bot is dead and has a termination_task_id
                Kill the VM
        3. Bot is dead and was last seen more than vm.Timeout ago
                Kill the VM

* All VMs are subject to a minPendingTime amnesty from this rule on
  their first boot (currently 10 minutes)

Change-Id: Ic6c6a3f96b9b7315fc427e6dd7bd2a6b7e7ce27d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6277626
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Gowri Denduluri <gowriden@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Anushruth Sasivehalli <anushruth@chromium.org>