[scheduler] Ignore pre-existing tags
Previously, adding a new scheduler trigger or changing on existing
trigger's regex caused the trigger to launch a job for every existing
tag that matched the regex, of which there could be thousands or more,
which would quickly overwhelm a team's infrastructure. Generally the
intent when setting up a trigger that matches tags is for the trigger to
only launch jobs for *new* tags, not all historical ones.

We still trigger on pre-existing branches, since it's much less likely
that a trigger would match many different branches, and more likely that
someone setting up a new job would actually want the job to trigger
immediately on whatever the latest commit on the branch was.

Likewise, we still trigger if an existing tag is changed to point to a
different commit, which sometimes happens despite being problematic.

Fixed: 382512264
Change-Id: I8baed9efd6b4e05bc8d773a5ca593ca3315555e6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6278797
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>