[swarming] Implement /bot/oauth_token endpoint for "task" accounts.
There are 4 similar flows depending on combination of
account_id (either "task" or "system") and what kind of token
is requested ("oauth_token" or "id_token").

This CL implements "task" + "oauth_token", but it was extracted
from a larger CL that implements all 4. For that reason it is
structured somewhat weirdly (with an interface in the middle).
This will make more sense when the remaining 3 cases are
implemented.

R=chanli@chromium.org
BUG=b/355013054

Change-Id: I7607b6136741a7737e64adba57efaf9a082b4418
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6277640
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>