cv: searching reusable tryjob in bb also populates Tryjob.CLPatchsets
Before, CLPatchsets computed from the CLs of the Run are used. This is
not 100% correct because Tryjob.CLPatchsets is supposed to reflect the
CL and patchset information associated with the Tryjob (Buildbucket
build) and it may not be the same as the one computed from the Run due
to equivalent patchset.

This is also needed to decide whether Tryjob can be reused by checking
if a footer value has been added/removed/changed between the patchset of
the Tryjob and the patchset of the Run.

Bug: 40785878
Change-Id: I40490624d8d6798a7dffdda14de0599f5a228cb6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6255435
Auto-Submit: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Oliver Newman <olivernewman@google.com>