[cv] Invalidate previous tryjobs on certain footer changes
This change introduces a new field to the `Tryjob` message of
commit-queue.cfg, "disable_reuse_footers". If a CL has any of those
footers added, removed, or their values changed, then previous runs of
the given tryjob will not be reused on the new patchset of the CL.

This is useful because some recipes read commit message footers and
alter their behavior based on them, so CQ results on otherwise-identical
patchsets that have different footer values are not comparable.

For example, the Fuchsia project skips unaffected tests by default in
presubmit, but supports a commit message footer, `Run-All-Tests: true`,
to override this behavior. Consider the following situation:

1. Engineer uploads a CL without the footer set.
2. Does a CV dry run on the CL.
3. Uploads a new patchset with no code changes, but with `Run-All-Tests:
   true` set.
4. Does a CV full run (less than 24 hours after the dry run).

In this case, the engineer expects that CV will re-run all necessary
builders, respecting the footer, in #4 before submitting the CL. But
prior to this change, CV would reuse the previous results, and submit
the change without running all tests.

Now, as long as the project has `Run-All-Tests` listed in
`disable_reuse_footers` for the necessary tryjobs, CV will behave as the
user intends and rerun the tryjobs.

Bug: 40785878
Bug: 377534181
Change-Id: I4371c530cc6705d8c1695b5d4f26786e04bd1402
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6242582
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>
Auto-Submit: Oliver Newman <olivernewman@google.com>