[lucicfg] Add support for disable_reuse_footers
For any commit message footers listed in the `disable_reuse_footers`
parameter to `luci.cq_tryjob_verifier()`, CV now checks the footer
value in the previous and current patchset when determining whether to
reuse a previous tryjob. If the value of any of the listed footers has
changed, or if any of the footers have been added/removed, then CV will
not reuse the previous tryjob even if the patchsets are otherwise
equivalent.

Bug: 40785878
Bug: 377534181
Change-Id: I0dd1104a3450fbfa2856cbf3f4ae2033bd7a8bca
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6244343
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>