[Fleet Console] Show the device table as soon as data is loaded
Here the thing is that GetDeviceDimensions call takes around 2-3s
because of that overall the devices first page takes longer to load. As
we basically have all the data to show for the first page as
soon as we get the data from the ListDevices (which is fast), we can
move the loading state to the filters and columns instead of the whole
page.

https://screencast.googleplex.com/cast/NDgzNTU3MjAwNTczMjM1Mnw3NDc0NTM0NS0xNg

Bug: b/398005480
Change-Id: Ic4477a60ce2f859242053d406fd031988e759b14
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6299150
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Vaghinak Vardanyan <vaghinak@google.com>