[Fleet Console] Fix device table pagination issue
Changing the page always results on getting back to the first page. The
reason is that I put the pagerCtx into the useEffect's dep array as per
linter's exhaustive dep check. And every time it got changed this
useEffect caused it to be reset.

Bug: b/398005480
Change-Id: I30b95b38316d8e4780440e40747d9e5ddbfc2230
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6298235
Reviewed-by: Pietro Scuttari <pietroscutta@google.com>
Commit-Queue: Vaghinak Vardanyan <vaghinak@google.com>