[cipd] The client side of the CIPD proxy "protocol".
Basically, if there's CIPD_PROXY_URL env var, send all
HTTP requests to it. The requests themselves are not
changing, i.e. they still use the original `Host` header:
this is how the proxy would know what CIPD server to hit.

This is an URL (and not just a domain socket path) to potentially
allow "more remote" proxies in the future (e.g. a proxy in the
local network that dedups GCS downloads). There are no plans to
pursues this, but it would be awkward to have two different proxy
mechanisms and URL formats in case we want to in the future.

Note that we need to turn off authentication
(since it also hits network).

R=iannucci@chromium.org
BUG=b/397997317

Change-Id: I3a53789acde1aa4031a83d675b01c150d950e434
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6293845
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>