[errors] Begin dismantling bespoke error annotation system.
This introduces a new package, errtag, which is very nearly a drop-in
replacement for the existing errors tagging functionality, except that
it's built exclusively on the stdlib errors package and follows modern
conventions on how this sort of thing should be done (while making some
concessions for the existing codebase to minimize the amount of
widespread code changes which all need to happen "at once").

The new library is also typesafe, meaning it's no longer possible to
have type confusion around the value associated with a given tag.

R=gregorynisbet, vadimsh

Change-Id: I4d263a5d669d401d7d3cff1839908e58950d7533
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6242995
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Robbie Iannucci <iannucci@google.com>