[Analysis] Move and rename TestVariantIdentifier proto.
In preparation of adding a copy of ResultDB's proto. A reference
to this proto is probably in ChromeOS's recipes so we will have
to do a manual roll.

This does not affect the parsing or encoding of RPC requests,
i.e. should not break existing clients.

BUG=b:398112066

Change-Id: Ia036406a2020df40c4301d318ea64d656ff0fadd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6312274
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>