[ResultDB] Redefine approach for exposing schemas to clients.
Replace the list method with a new get method that is simpler to
use and is more extensible if we have additional information
about test schema that clients need to query, such as test status
labels.

To avoid exceeding the limit for a response, we add a validation
that the total size of all schemes configured for a ResultDB
deployment may not exceed 1000 items or 100KB (whichever is less).

BUG=b:398112066

Change-Id: I3a30b371c3c78feff59f13caffb85b466f5e7317
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6320436
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>