Fix horizontal spacing of console headers
I'm working on improving the header and console groups that appear in
it. This is leading to the "links" and "Oncalls" to take up more of
the available horizontal space. See:
https://ci.chromium.org/p/chromium/g/chromium.gpu%20rotation/console
To remedy this, apply a min width to the last console-header-column
since both sides of this table use the same type.

Bug: 40873502
Change-Id: Ib8696de9beb454afa957d90f0d85d90d5bcb9539
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6325391
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>