Fix secondary filter menu positioning
demo:
https://17840-7565de0-pietro-fix-secondary-menu-posit-dot-luci-milo-dev.appspot.com/ui/fleet/labs/devices

You can still get into weird states but I think it's an improvement.
https://screenshot.googleplex.com/5FFm4qpbqUh8kGM.png

The reason is that the position of the menu depends on the
width of the primary filter menu. When you type a character in the
search box, if the new sub-menu position gets calculated before the
browser has finished repainting the primary menu the resulting position will be wrong. So essentially we have a race condition between the browser repainting the primary menu and the secondary menu position being calculated.
I tried forcing the calculation to happen after the
browser repaint but I was not able to find a good solution.

Bug: b/398007448
Change-Id: Iacc3d9b9c1c7af2607095281d629212f4dfb6fde
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6330403
Reviewed-by: Konrad Polarczyk <polarczyk@google.com>
Commit-Queue: Pietro Scuttari <pietroscutta@google.com>
Auto-Submit: Pietro Scuttari <pietroscutta@google.com>