[lucicfg] Introduce a notion of "entry point package".
It is a representation of the main package with all its dependencies
resolved and prefetched. It will be the outcome of the process of
loading PACKAGE.star and/or PACKAGE.lock files.

It can be loaded from disk or (in the future) from some other
source. For now use the compatibility logic: each entry point
resides in a root of a package with no other dependencies.

There's one complication: the previous version of this code
used absolute path to the package root for some of the logic
of populating project.cfg.

We no longer assume a disk checkout. Use a package path relative
to the repository root instead (which always makes sense, even
in absence of a checkout). This necessitates restricting lucicfg
output config dir to be under to the same repo that contains
the config file. This should not be an issue in practice.

The relevant code was super confusing. Attempt to comment it better
and add some tests.

R=iannucci@chromium.org, fancl@chromium.org

Change-Id: I740b739d9e8bffdb45b61e12283fb55bb551c7fd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6337414
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@google.com>