[ResultDB] Define treatment of variable depth test hierarchies.
Reserving the '/' separator and enforcing some sensible usage via
additional validation will create opportunity to add UI sugar for test
cases with additional hierarchy in future.

At the same time, the test hierarchy design was clear in that
we believed the (up to) four component hierarchy is enough for the
vast majority of usages and the complexity of full support of variable depth was not worth it. We want to be able to have a fixed set of
fields to map to primary keys and don't want to be doing substring
manipulation in SQL.

As per go/resultdb-subtest-support-evaluation.

BUG=b:398112066
TEST=None

Change-Id: Ie7024441d94d28fef12c9becafa2a215e1b372fd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6346098
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>