[prpc] Show that we are waiting for a message on STDIN when verbose is enabled.
Make the behavior of prpc a little bit more discoverable in verbose mode. Right now the command
below will silently wait forever.

A user's first instinct is to kill it and try it again with verbose mode, so we should show that
we're waiting for a message on stdin when that happens.

$ prpc call localhost:8800 servicerpc.service ListThings '{}'

Change-Id: I734efd03f302479e95d5cce8b4cf1514b3d6a5c0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6362957
Reviewed-by: Justin Suen <justinsuen@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>