[lucicfg] A graph structure used during dependency traversal.
It is a directed graph where nodes are `(package, version)` pairs
(e.g. concrete versions of packages) and edges represent direct
dependencies between such versions.

Edges are annotated (this will be used later to produce rich
error messages).

Keeps track of all observed version of packages, to be able to
pick the most recent one later (but this will be done externally,
since we'd want to do that in parallel and shoving this machinery
into this graph struct felt forced).

The graph can be built in parallel, finalized, and the traversed
in parallel. Traversal doesn't have to follow the edges (this will
be useful when traversing along "most recent version" path).

R=iannucci@chromium.org

Change-Id: I5f05347f3a6715f4b935cf76f2c86f9d9e6d7847
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6369620
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>