[lucicfg] Add some abstractions for working with remote deps.
The core abstraction is Repo interface, which will eventually
be implemented on top of depsource APIs. It will also be
mocked in tests to allow to use local files as substitute for
"remote" repos.

This also attempts to explain the edge case related to
remote dependencies depending back on the local dependencies,
since its consequences leak into the API (we need a special
version tag to represented "some uncomparable local revision"
that should always compare to itself in MVS and never to
real revisions).

R=iannucci@chromium.org

Change-Id: I8b85dc2be8c32a83085d81724ec7f4c4aac96e9e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6370968
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>