[swarming] Update format of the generated effective RBE bot ID.
Looks like RBE doesn't mind ":" in the bot ID (in fact, I could not
find any mentions of it being validated in any way, not in docs, not
in the code). I tried running a bot with ":" in the ID and it worked.

We can use it to guarantee generated IDs will never clash with real
IDs (since real IDs can't have ":" in them). The previously used
"--" separator was problematic, because it was also already used by
"composite" bot IDs, like "build4-h9--device1", so there was a chance
of collision.

Also put the dimension name there while at it to make such bot IDs
slightly less mysterious when they are noticed in logs.

Finally, do not confuse the generated ID with the real ID when
processing RBE PubSub notifications.

R=chanli@chromium.org
BUG=b/355013224

Change-Id: I90f8ccb1b2ad568265e68ec6443e572bd78db73d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6393921
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>