[lucicfg] Remove special meaning of "@__main__" and "@stdlib".
Now the main package in lucicfg is known by the name it was
defined as in its PACKAGE.star. It will remove some edge cases
and make "executable-like" and "library-like" packages more
similar.

Retain "@__main__" for legacy lucicfg code (that doesn't have
PACKAGE.star).

The only downside is that we now need to know what is the
main package name to stringify ModuleKey in a user friendly
form. There's now a new method for this which takes
*starlark.Thread.

Intentionally remove Interpreter.Init to make sure all
existing code that uses it will stop compiling (since it
now needs to be updated to call LoadBuiltins).

Also modernize terminology in interpreter.go comments.

R=iannucci@chromium.org

Change-Id: I12b5f4c8dd049413114228de46d4ab7d8a4b3787
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6378863
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>