[lucicfg] Remove pkg.source.submodule(...) for now.
It brings a variety of problems related to generation of
the lockfile. It can be added later if actually needed.

To stop possible attempts to load submodule'd dependencies
as local, verify that pkg.source.local(...) actually points
to a directory in the same repository as the root (should
have done that anyway, just forgot). This is actually not
trivial if we consider transitive local dependencies
(because the mechanism that loads transitive dependencies
doesn't know they are local, requiring to extend
Repo abstraction a little bit).

R=iannucci@chromium.org

Change-Id: I545ecce3dfbcda167592c0feb0e94dabae59dbe2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6378885
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>