[ResultSink] Implement -shorten-ids option.
This option activates automatic shortening of test IDs to 350
bytes (prior to being combined with Test ID prefix / module).

The logic to do this correctly is non-trivial in case of structured
test IDs, so we'd rather implement it centrally for those who
need it, even if we don't encourage its use.

This CL also fixes the handling of the empty test_id, which is
allowed where test ID prefix is set.

BUG=b:398112066
TEST=Integration tests

Change-Id: Iea190d7990ec44a05016fab1270bcdb1f61a0f44
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6420115
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Beining Chen <beining@google.com>
Feels: Beining Chen <beining@google.com>