[swarming] Refactor how BotInfoUpdate.TaskInfo works.
Unlike most other BotInfo fields, a TaskID assigned to
a bot can't be changed arbitrarily. There's essentially
a state machine that restricts how tasks and bots work
(otherwise the bot or the task would end up in a weird
state).

This state machine transitions are now (kind of) enforced
by the BotInfoUpdate, in particular even if tests are mocking
it (as long as they still call PanicIfInvalid).

This will be relied upon in the following CL that extended
this state machine (kind of) to emit "task was abandoned"
events if the bot dies unexpectedly while working on a task.

R=chanli@chromium.org
BUG=b/355013593

Change-Id: Ie5308a6e342344b5433d3d8cec6f87c1a6092c7a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6423712
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>