[swarming] Make BotInfoUpdate recognize abandoned tasks.
A task gets abandoned by a bot if it is assigned to the bot
when this bot disappears (either becomes missing, gets deleted
or shutdown), or if it is suddenly reports itself as idle (e.g.
reconnects after unexpected reboot) or working on another task
(should not really be happening...).

To preserve existing quirky behavior when dead bots are listed
with TaskID's still set to whatever they were running when they
died, BotEventMissing event is NOT resetting TaskID. But it
still should result in the task being abandoned. This forces
to introduce a new BotInfo.LastAbandonedTask field.

R=chanli@chromium.org
BUG=b/355013593

Change-Id: I37eede2b84323601d1b128c002f65cd97b81c9eb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6423619
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>