[gitsource] Add fetcher implementation.
Adds the main 'fetcher' implementation. This will be used as the
primary interface for obtaining blob data from the remote.

This implements multiple prefetch optimizations - one which will
fire based on a callback which can process the tree of the target
commit to pull objects under the target subdirectory which look like
the right files to get, and a second where we can specifically
ensure that a batch of arbitrary objects is cached.

R=fancl, fancl@chromium.org, vadimsh@chromium.org

Bug: 367699933
Change-Id: I21ef7e588939675e85babfdd499076c4254a215c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6313604
Auto-Submit: Robbie Iannucci <iannucci@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@google.com>