[gitsource] Add starwalk demo/manual testing helper.
This implements a `starwalk` tool which is able to use the gitsource
implementation to pull down and cache all the .star files in
a target repo/ref@commit and lay them out on disk.

For extra flair, this also parses the incoming star files to see if
any of them are lucicfg containing `io.read_file` or `io.read_proto`
functions which take literal string arguments and fetches them, too.

The real lucicfg implementation will just use the Fetcher
implementation directly in memory as part of the starlark
evaluation, and won't need/want to put these down on disk, but this
tool is/was useful for validating the gitsource implementation as
basically functional during its developement.

R=fancl, fancl@chromium.org, vadimsh@chromium.org

Bug: 367699933
Change-Id: I63aacfe1222a763e656a4a450623fc11b02348c0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6315858
Commit-Queue: Robbie Iannucci <iannucci@google.com>
Auto-Submit: Robbie Iannucci <iannucci@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>